working on poast dual regression
