+++ pyproject.toml
@@ -39,3 +39,4 @@ nfact_config = "NFACT.config.__main__:nfact_config_main"
 nfact_dr = "NFACT.dual_reg.__main__:nfact_dr_main"
 nfact = "NFACT.pipeline.__main__:nfact_pipeline_main"
 nfact_Qc = "NFACT.qc.__main__:nfactQc_main"
+nfact_PD = "NFACT.post_dr.__main__:nfactPD_main"
